+++ src/main/java/warmup/one/WarmupOneFifthSet.java
@@ -14,4 +14,15 @@ public class WarmupOneFifthSet {
 
         return character + str + character;
     }
+
+    /**
+     * Return true if the given non-negative number is a multiple of 3 or a multiple of 5. Use the
+     * % "mod" operator.
+     *
+     * @param n
+     * @return boolean
+     */
+    public boolean or35(int n) {
+        return n % 3 == 0 || n % 5 == 0;
+    }
 }