+++ src/main/java/warmup/one/WarmupOneSixthSet.java
@@ -1,0 +1,12 @@
+package warmup.one;
+
+public class WarmupOneSixthSet {
+    /**
+     * Given a string, return true if the string starts with "hi" and false otherwise.
+     * @param str
+     * @return
+     */
+    public boolean startHi(String str) {
+        return str.startsWith("hi");
+    }
+}